Improve look and behaviour of source editing page.
This commit contains several changes. First, the CodeMirror editor is now
resizable, which is particularly useful for the tutorial page. The resizing
requires jQuery UI, hence why that plugin has been added.

Second, the 'view/edit soure' link is no longer shown for the source editing
page, since much of that code is fairly complicated an not particularly useful
for newcomers to the framework.

Third, I've change the markup behind the 'View' and 'Controller' tab buttons
to be more standard and to have a greater differentiation between the active
and non-active tabs.